chore(form): simplify and optimize conditional callback resolution a little bit
