+++ packages/sanity/src/form/store/conditional-property/index.ts
@@ -1,1 +1,1 @@
-export * from './resolveConditionalProperties'
+export * from './callConditionalProperty'
